backend/requirements.txt
fastapi
uvicorn[standard]
duckdb
pandas
numpy
google-genai
pydantic
pydantic-settings
python-dotenv
